+++ core/song/lexer/lex/lexingFunctions/LexNewLine.go
@@ -7,7 +7,7 @@ import (
 
 // LexNewLine
 func LexNewLine(lexer *lex.Lexer) lex.LexingFunction {
-	lexer.Position += len(lexertoken.NEWLINE)
+	lexer.Position += uint(len(lexertoken.NEWLINE))
 	lexer.PushToken(lexertoken.TOKEN_NEWLINE)
 	return LexSection
 }
