+++ core/song/lexer/lex/lexingFunctions/LexParenthesis.go
@@ -7,14 +7,14 @@ import (
 
 // LexLeftParenthesis
 func LexLeftParenthesis(lexer *lex.Lexer) lex.LexingFunction {
-	lexer.Position += len(lexertoken.LEFT_PARENTHESIS)
+	lexer.Position += uint(len(lexertoken.LEFT_PARENTHESIS))
 	lexer.PushToken(lexertoken.TOKEN_LEFT_PARENTHESIS)
 	return LexWhisperedLyrics
 }
 
 // LexRightParenthesis
 func LexRightParenthesis(lexer *lex.Lexer) lex.LexingFunction {
-	lexer.Position += len(lexertoken.RIGHT_PARENTHESIS)
+	lexer.Position += uint(len(lexertoken.RIGHT_PARENTHESIS))
 	lexer.PushToken(lexertoken.TOKEN_RIGHT_PARENTHESIS)
 	return LexSongLine
 }
